+++ MyohanMeeting/src/main/java/meet/myo/domain/cat/Neutered.java
@@ -1,4 +1,4 @@
-package meet.myo.domain;
+package meet.myo.domain.cat;
 
 public enum Neutered {
     NEUTERED, UNNEUTERED, UNKNOWN
